Enforce owner permissions on views.
Non-owners should be able to see their own data. We probably want to
think about a better way about dealing with permissions/privacy to make
sure that a bug that leaks data is not introduced in the future, and to
allow for use-cases where we do want to users to be able to share data.