Create ref fasta index before kicking off concurrent alignment jobs.
Also update the interface to align_with_bwa() to accept an existing
ExperimentSampleToAlignmentGroup object for performing a re-alignment.

This commit also includes a script_debugger.py module that will be a
store of test calls to existing methods, along with calls to set up the
Django environment and everything. I find myself doing this a lot so
might as well keep this hackpad around.