Added snpeff build and signals (no vcf annotation support yet).
Now signals.py looks for new reference genomes and also runs jbrowse and snpeff prep when a new genome sequence file (genbank or fasta) is added. Next up is to add snpeff vcf annotation support to the alignment pipeline.

I also changed the name of the prepare_reference_genome function in jbrowse_util to include jbrowse in the name, so that it would be clearer when used in other functions/contexts.