Support add/remove to Sets for melted view.
There are several hacks in this implementation, but I'm committing in
order to move forward and be able to nail down exact behavior before
going back and building it correctly. I've opened a bug to nail down
behavior in #53.

Since, the basics of add/remove in melted view works, I'm marking the
related issue fixed #38.