VariantAlternate now populates a JSONField().
It works just like VariantCallerCommonData, except it uses only
per-alternate fields. All per-alternate fields that were in
VariantCallerCommonData and were dealt with separately are now
in VariantAlternate instead, and the flow is the same. get_per_alt_field()
is no longer necessary (though I haven't removed it yet).

One TODO: currently we are parsing VA fields on a per-sample basis,
but in many cases it will be useful to look at alternate for variants
that have no samples, like those imported from as a Variant Set. I'll
deal with this later.