Support alt_sample "!=" case in Variant filtering.
The != case reveals that we can only do the initial SQL filter by values
that are sample-agnostic, and then we have to do the per-sample
separately. An example that illustrates this is the query "alt_value !=
'A'" will actually return no Variants on the initial SQL query if there
exist any alt_value other than A.  Instead, we have to handle this
filter key on a per-sample basis.