Test for the full alignment and variant calling pipeline.
This runs the full pipeline on pre-fabricated data where we know which
variants were introduced.

This passes with both bwa and bwa_mem. I wrote this while trying to get
to the bottom of issue #93, where it's not clear, but it seems like
there is something strange about bwa_mem. This test does not reveal
any discrepancy between the two.

Dan suggests that perhaps the issue observed there was actually with
JBrowse. Continuing to debug.