Point jsonfield to dev version that works with Postgresql 9.3.
After updating to Postgresql 9.3, I was getting failing tests because
the jsonfield package was trying to use the native Postgres 'json' type,
but something was failing. A close-enough issue is here:
https://github.com/bradjasper/django-jsonfield/issues/57. Until this is
fixed, I'm pointing to a fork that provides a simple fix for this where
we go back to pre-Postgres 9.2 behavior pretending we don't have a
special json field.