Make variant set semantics work.
Before this commit, our materialized view was omitting variant set
associations that didn't have an ExperimentSample association.

The solution to getting this to work was to use a UNION in the
materialized view SQL statement. Specifically, we make sure we get all
variant to set relationships by ignoring samples, and then UNIONing with
the joined table that does include the relations.