Celery tasks not being registered following DBG refactor.
Django celery only finds tasks that are in modules called tasks.py. In
this case, read_alignment was being found because it is imported from in
signals.py - I couldnt find a reference on why this is. Upon explicitly
declaring the import, there turned out to be a circular import because
importing from models/__init__.py causes an import of signals.py which
caused the circular import. To solve the circular import, I broke out
the functions that signals.py needs from read_alignment into a new
module read_alignment_util.

We should consider upgrading to Celery 3.1 which is much better at
auto-registering tasks.