Fix setup.py so that it works from commandline.
The fix required removing the import of InputError. The problem with
importing that is it it's located in the main package, and importing
something from the main package calls main/__init__.py to run, which
imports signals.py, which requires the Django environment to be setup.

We probably need to rework setup.py in general though.