IntegrationTestSuiteRunner. Doesn't quite work yet.
Also added a script that runs integration tests.

This starts up celery that talks to the test database. However, I
haven't implemented a good way to block on running asserts until the
test task is done, so right now certain tests are failing
because they aren't waiting for async tasks to complete.