Connect djcelery_testworker WorkerOutputThread to console.
Right now, if there's an error in celery, the console will hang,
requiring a ctrl+c to move past it. Where we really want to get is to
have behavior like in typical tests, where the the test process will
end on its own and dump the relevant error output for debug. So the
question is why does the celery process failing result in the test
hanging?