Prevent erroneous form submit on 'enter' keypress.
Most of our forms expect to be programmatically populated in part
(including csrf) and submitted. Pressing enter goes with the default
browser (or Bootstrap?) behavior of submitting the form. For now,
prevent that from happening.

However, there are places where it makes sense for the enter key to
work, but correctly, so there may be a better solution.

Fixes #306.