Added option to vcf_parser for skipping het-only variants.
If a variant has only GT_TYPE=1 and GT_TYPE=2 samples, then
it is probably not work adding to the database by default,
since it is likely the result of mapping errors.

The next step is to allow the user to turn this on and off.
For now you can pass skip_het_only=True to parse_vcf()
and parse_alignment_group_vcf() in order to turn on
het-only skipping. I also added a new test for this option.