Replaced GATK CallableLoci with custom pysam-based function.
Settings for the new code were added to conf.global_settings.py and
a test was added in test_read_alignment. For it to work well, we
need to use the `-a` flag `in bwa_mem`, but before that can work
we have to adjust how the de-novo alignment sam preprocessing works.

The CallableLoci settings might have to be tweaked when we look at
real data.