Configure travis-ci.
Temporarily commenting out optmage_util test until I get hybrid-ss-min
in.

Note that this still leaves one test failing, which is due to the known
bug in the previous pull request merge which we'll fix in a follow up
commit so we can get the satisfying feeling of seeing the build go from
failing to success!