Update requirements and fix travis-ci build.
Travis-ci.org updated their test machiens to use 2.7.13, which broke our
build. I was pretty certain this was due to out-dated dependency issues.

We were getting a failure during install of the `distribute` python
dependency related to `import _markerlib` failing. The Internet suggests
this is a remnant of `distribute` being deprecated with the latest
version of pip so the best fix would be to get rid of this dependency.
Turns out it was coming from PyVCF==0.6.4. The latest PyVCF==0.6.8
breaks our tests due to some interface changes. However, 0.6.7 works.
But 0.6.7 still has an unnecessary dependency on `distribute` which can
be deleted without affecting build. Thus we created a churchlab fork of the
PyVCF repo which is then reset to release 0.6.7 but with an extra commit
that removed the `distribute` dependency.

Fixing the build also required upgrading six. I just upgraded to the
latest version 1.10.0, which passes all tests.